feat/split-e2e-tests: set 8 workers
